ipq40xx: qca8k-ipq4019: fixup RGMII support
Previously, just the generic RGMII support was added.
However this is not enough as almost all devices will require some kind
of delays, be it on RX, TX or both.
Since the IPQ40xx SoC cannot add these itself, it relies on the PHY-s to
do so depending on the design, and that requires supporting all of the
RGMII PHY modes.

Since IPQ40xx itself doesnt need to do any kind of special setup for ID,
RXID and TXID allow setting these so that Phylink can propagate them to
the PHY driver so it sets correct delays.

Signed-off-by: Robert Marko <robert.marko@sartura.hr>